feat(experimental-utils): Add types for eslint-scope
