+++ packages/experimental-utils/src/index.ts
@@ -1,7 +1,8 @@
 import * as ESLintUtils from './eslint-utils';
 import * as TSESLint from './ts-eslint';
+import * as TSESLintScope from './ts-eslint-scope';
 
-export { ESLintUtils, TSESLint };
+export { ESLintUtils, TSESLint, TSESLintScope };
 
 // for convenience's sake - export the types directly from here so consumers
 // don't need to reference/install both packages in their code
