+++ packages/experimental-utils/src/ts-eslint-scope/Options.ts
@@ -1,0 +1,21 @@
+import { TSESTree } from '@typescript-eslint/typescript-estree';
+
+type PatternVisitorCallback = (
+  pattern: TSESTree.Identifier,
+  info: {
+    rest: boolean;
+    topLevel: boolean;
+    assignments: TSESTree.AssignmentPattern[];
+  },
+) => void;
+
+interface PatternVisitorOptions {
+  processRightHandNodes?: boolean;
+}
+
+interface Visitor {
+  visitChildren<T extends TSESTree.BaseNode | undefined | null>(node?: T): void;
+  visit<T extends TSESTree.BaseNode | undefined | null>(node?: T): void;
+}
+
+export { PatternVisitorCallback, PatternVisitorOptions, Visitor };
