+++ packages/experimental-utils/src/ts-eslint-scope/index.ts
@@ -1,0 +1,12 @@
+import { version as ESLintVersion } from 'eslint-scope';
+
+export * from './analyze';
+export * from './Definition';
+export * from './Options';
+export * from './PatternVisitor';
+export * from './Reference';
+export * from './Referencer';
+export * from './Scope';
+export * from './ScopeManager';
+export * from './Variable';
+export const version: string = ESLintVersion;
