+++ packages/parser/typings/eslint.d.ts
@@ -1,0 +1,14 @@
+declare module 'eslint/lib/util/traverser' {
+  import { TSESTree } from '@typescript-eslint/experimental-utils';
+  const traverser: {
+    traverse(
+      node: TSESTree.Node,
+      options: {
+        enter?: (node: TSESTree.Node, parent: TSESTree.Node) => void;
+        leave?: (node: TSESTree.Node, parent: TSESTree.Node) => void;
+        visitorKeys?: Record<string, string[]>;
+      },
+    ): void;
+  };
+  export = traverser;
+}
